+++ stories/NumericInput.stories.js
@@ -32,10 +32,10 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - label',
+  .add('label',
     withState({ value: '' }, store => (
       <NumericInput
-        label="Some label"
+        label="Amount"
         value={store.state.value}
         maxBeforeDot={6}
         maxAfterDot={6}
@@ -45,7 +45,7 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - placeholder',
+  .add('placeholder',
     withState({ value: '' }, store => (
       <NumericInput
         value={store.state.value}
@@ -58,22 +58,36 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - focus / blur',
-    withState({ value: '', focused: false, blurred: false }, store => (
+  .add('autoFocus',
+    withState({ value: '' }, store => (
       <NumericInput
+        autoFocus
         value={store.state.value}
-        placeholder="focus / blur"
+        placeholder="18.000000"
         maxBeforeDot={6}
         maxAfterDot={6}
         onChange={value => store.set({ value })}
-        onFocus={() => store.set({ focused: true })}
-        onBlur={() => store.set({ blurred: true })}
         skin={InputSkin}
       />
     ))
   )
 
-  .add('send amount - error',
+  .add('onFocus / onBlur',
+    withState({ value: '', focused: false, blurred: false }, store => (
+      <NumericInput
+        value={store.state.value}
+        placeholder="onFocus / onBlur"
+        maxBeforeDot={6}
+        maxAfterDot={6}
+        onChange={value => store.set({ value })}
+        onFocus={() => store.set({ focused: true, blurred: false })}
+        onBlur={() => store.set({ blurred: true, focused: false })}
+        skin={InputSkin}
+      />
+    ))
+  )
+
+  .add('with error',
     withState({ value: '' }, store => (
       <NumericInput
         label="Amount"
@@ -86,7 +100,7 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - beforeDot(3) and afterDot(4)',
+  .add('beforeDot(3) and afterDot(4)',
     withState({ value: '' }, store => (
       <NumericInput
         label="Amount"
@@ -100,7 +114,7 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - maxValue(30000) - unenforced',
+  .add('maxValue(30000) - unenforced',
     withState({ value: '' }, store => (
       <NumericInput
         label="Amount"
@@ -113,7 +127,7 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - minValue(50) - unenforced',
+  .add('minValue(50) - unenforced',
     withState({ value: '' }, store => (
       <NumericInput
         label="Amount"
@@ -127,7 +141,7 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - maxValue(30000), minValue(50) - unenforced',
+  .add('maxValue(30000), minValue(50) - unenforced',
     withState({ value: '' }, store => (
       <NumericInput
         label="Amount"
@@ -142,7 +156,7 @@ storiesOf('NumericInput', module)
     ))
   )
 
-  .add('send amount - maxValue(30000), minValue(50), enforceMax=true, enforceMin=true',
+  .add('maxValue(30000), minValue(50), enforceMax=true, enforceMin=true',
     withState({ value: '' }, store => (
       <NumericInput
         label="Amount"
@@ -153,22 +167,6 @@ storiesOf('NumericInput', module)
         maxAfterDot={6}
         enforceMax
         enforceMin
-        onChange={value => store.set({ value })}
-        skin={InputSkin}
-      />
-    ))
-  )
-
-  .add('send amount - onChange',
-    withState({ value: '' }, store => (
-      <NumericInput
-        label="Amount"
-        value={store.state.value}
-        placeholder="0.000000"
-        maxBeforeDot={12}
-        maxAfterDot={6}
-        maxValue={45000000000}
-        minValue={0.000001}
         onChange={value => store.set({ value })}
         skin={InputSkin}
       />