+++ stories/Select.stories.js
@@ -5,11 +5,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withState } from '@dump247/storybook-state';
 
-// components
-import { FormField, Select } from '../source/components';
-
-// skins
-import { SelectSkin, FormFieldSkin } from '../source/skins/simple';
+// components & skins
+import { Select } from '../source/components';
+import { SelectSkin } from '../source/skins/simple';
 
 // themes
 import SimpleTheme from '../source/themes/simple';
@@ -51,7 +49,7 @@ const COUNTRIES_WITH_DISABLED_OPTIONS = [
 storiesOf('Select', module)
   // ====== Stories ======
 
-  .add('countries - options',
+  .add('options',
     withState({ value: '' }, store => (
       <Select
         value={store.state.value}
@@ -62,25 +60,19 @@ storiesOf('Select', module)
     ))
   )
 
-  .add('countries - label',
+  .add('label',
     withState({ value: '' }, store => (
-      <FormField
-        label="Some label"
-        skin={FormFieldSkin}
-        render={() => (
-          <Select
-            value={store.state.value}
-            onChange={value => store.set({ value })}
-            label="Countries"
-            options={COUNTRIES}
-            skin={SelectSkin}
-          />
-        )}
+      <Select
+        label="Select a country"
+        value={store.state.value}
+        onChange={value => store.set({ value })}
+        options={COUNTRIES}
+        skin={SelectSkin}
       />
     ))
   )
 
-  .add('countries - placeholder',
+  .add('placeholder',
     withState({ value: '' }, store => (
       <Select
         value={store.state.value}
@@ -92,7 +84,7 @@ storiesOf('Select', module)
     ))
   )
 
-  .add('countries - value',
+  .add('value',
     withState({ value: COUNTRIES[0].value }, store => (
       <Select
         value={store.state.value}
@@ -103,26 +95,20 @@ storiesOf('Select', module)
     ))
   )
 
-  .add('countries - error',
+  .add('with error',
     withState({ value: COUNTRIES[0].value }, store => (
-      <FormField
+      <Select
         label="Countries"
         error="You picked the wrong country"
-        skin={FormFieldSkin}
-        render={props => (
-          <Select
-            {...props}
-            value={store.state.value}
-            onChange={value => store.set({ value })}
-            options={COUNTRIES}
-            skin={SelectSkin}
-          />
-        )}
+        value={store.state.value}
+        onChange={value => store.set({ value })}
+        options={COUNTRIES}
+        skin={SelectSkin}
       />
     ))
   )
 
-  .add('countries - custom options template',
+  .add('custom options template',
     withState({ value: '' }, store => (
       <Select
         value={store.state.value}
@@ -139,27 +125,22 @@ storiesOf('Select', module)
     ))
   )
 
-  .add('countries - isOpeningUpward',
+  .add('isOpeningUpward',
     withState({ value: '' }, store => (
-      <FormField
+      <Select
+        isOpeningUpward
         className={styles.customMargin}
         label="Countries (opening upward)"
-        skin={FormFieldSkin}
-        render={() => (
-          <Select
-            value={store.state.value}
-            onChange={value => store.set({ value })}
-            options={COUNTRIES}
-            placeholder="Select your country …"
-            skin={SelectSkin}
-            isOpeningUpward
-          />
-        )}
+        placeholder="Select your country …"
+        value={store.state.value}
+        onChange={value => store.set({ value })}
+        options={COUNTRIES}
+        skin={SelectSkin}
       />
     ))
   )
 
-  .add('countries - with disabled options',
+  .add('with disabled options',
     withState({ value: '' }, store => (
       <Select
         value={store.state.value}
@@ -172,7 +153,7 @@ storiesOf('Select', module)
     ))
   )
 
-  .add('countries - rtl support',
+  .add('rtl support',
     withState({ value: COUNTRIES[0].value }, store => (
       <div dir="rtl">
         <Select